+++ lib/llvm-backend/src/lib.rs
@@ -6,7 +6,10 @@
     unused_unsafe,
     unreachable_patterns
 )]
-#![cfg_attr(not(target_os = "windows"), deny(dead_code))]
+#![cfg_attr(
+    all(not(target_os = "windows"), not(target_arch = "aarch64")),
+    deny(dead_code)
+)]
 #![cfg_attr(nightly, feature(unwind_attributes))]
 #![doc(html_favicon_url = "https://wasmer.io/static/icons/favicon.ico")]
 #![doc(html_logo_url = "https://avatars3.githubusercontent.com/u/44205449?s=200&v=4")]